Refactor into HostFile struct and modules
